requirements.txt
flask
pytest
requests
beautifulsoup4
python-dotenv